Playing with hyper parameters. Also included a variable threshold.
Playing with the hyper parameters and trying to understand each of their effects on the network. One issue is where the network neurons inputs start of uniformly between [0.3, 07] which has a threshold which stops all the neurons from firing. However, this threshold is easily surpassed by the next input easily since the higher weights are now centered where the digits will most likely be. This stop other neurons from getting the chance to develop. Also, a '3' and '8' digits typically cover more area than a '1' would, making them harder to develop without easily getting overrode by another digit with more surface area. This is the reason behind the variable threshold. More plasticity ideas to come!

It should be remembered that the main goal is simplicity for hardware implementation. Microcode will be required for learning rules. This memory space will most likely need to be small, so complexity should be minimized. Also, bit count will need to be taken into consideration.