Hyperparam tuning. Bug fixes.
Now able to achieve an accuracy of about 85% using 200 neurons. Further investigation is needed on how to label the neurons. Most likely a post 'most likely to fire' counting scheme. Still not parallel and I hate this the most, but the idea works. This is mainly to be put into hardware, ie an FPGA.